!972 remove useless code from v2.2 * fix: remove useless code * fix: remove useless code * Merge branch 'master' into zqh-master * fix: remove useless code * fix: compile error for gcc 13

+++ kwdbts2/ts_engine/include/ts_engine_schema_manager.h
@@ -44,7 +44,7 @@ class TsEngineSchemaManager {
 
   KStatus Init(kwdbContext_p ctx);
 
-  KStatus CreateTable(kwdbContext_p ctx, const uint64_t& db_id, const KTableKey& table_id, roachpb::CreateTsTable* meta);
+  KStatus CreateTable(kwdbContext_p ctx, const uint32_t& db_id, const KTableKey& table_id, roachpb::CreateTsTable* meta);
 
   KStatus GetTableMetricSchema(kwdbContext_p ctx, TSTableID tbl_id, uint32_t version,
                                std::shared_ptr<MMapMetricsTable>* metric_schema);
